agent/requirements.txt
botocore
strands-agents
strands-agents-tools[rss]
strands-agents[otel]
bedrock-agentcore
aws-opentelemetry-distro